Change from os-conf test release to zookeeper-release
Thhe BOSH team made changes to the CLI to enhance the information
reported by commands such as `bosh vms`, and this broke one of the bosh
disaster recovery acceptance tests. This is due to a new status being
reported when there are no processes running on the VM. Since to better
validate the test scenario that broke we require a VM with running
processes, we changed the test release to a release that include running
jobs (zookeeper-release).

[#167575114](https://www.pivotaltracker.com/story/show/167575114)

Co-authored-by: Maya Rosecrance <mrosecrance@pivotal.io>