re-enable truncate-db-testcase (#132)
[#186014099]

when we moved the bdrats tests to runway, we added a jumpbox deployment
where all the bbr commands are executed from. This required deploying
a jumpbox via the available bosh. But since that jumpbox would be managed
by the director that is getting backed up, some tests had to be deactivated
because the test execution made the bosh unavailable to execute `bosh ssh`.

This updates the tests to have a configurable jumphost (with provided ssh key).
The actual jumphost is the opsman from the smith env that CI claims before
running the tests.

the option to deploy the jumpbox is kept in place in case of some downstream
issue with this PR so were able to switch back quickly.