speed improvement for pipeline.
[ #186014099 ]

a while ago we switched this pipeline from deploying a snowflake
bosh-deployment based env to a smith env. The smith env comes with
tas. That means the B-DRATS will backup the blobstore while it contains
all TAS blobs. Which increases runtime for this ci from <30m to >3h.

This cleans out the TAS deployment before starting drats to avoid copying
and verifying huge blobstore contents.