do not use shuttle (#133)
* do not use shuttle

920dd85e926664cf93a211f9356765b7c65d1ea6 added a feature where the jumphost could be set via the config file.
this means that using sshuttle is unnecessary as all commands are executed on the
jumpbox or via BOSH_ALL_PROXY

* speed improvement for pipeline.

[ #186014099 ]

a while ago we switched this pipeline from deploying a snowflake
bosh-deployment based env to a smith env. The smith env comes with
tas. That means the B-DRATS will backup the blobstore while it contains
all TAS blobs. Which increases runtime for this ci from <30m to >3h.

This cleans out the TAS deployment before starting drats to avoid copying
and verifying huge blobstore contents.